Added security to some routes

+++ src/controller/todo.py
@@ -3,6 +3,7 @@ from src.schema import schemas
 from sqlalchemy.orm import Session
 from src.model import models
 from src.model.db_conn import get_db
+from src.auth import auth
 
 router = APIRouter(
     prefix='/api/v1/todos',
@@ -10,39 +11,49 @@ router = APIRouter(
 )
 
 
-@router.get('/')
-def get_all_todo(db: Session = Depends(get_db)):
-    todos = db.query(models.Todos).all()
+@router.get('/', response_model=schemas.Todo)
+def get_all_todo(db: Session = Depends(get_db), user_id: int = Depends(auth.get_active_user)):
+    todos = db.query(models.Todos).filter(
+        models.Todos.owner_id == user_id).all()
     return todos
 
 
-@router.post('/', status_code=status.HTTP_201_CREATED)
-def create_post(todo: schemas.CreateTodo, db: Session = Depends(get_db)):
+@router.post('/', status_code=status.HTTP_201_CREATED, response_model=schemas.Todo)
+def create_post(todo: schemas.CreateTodo, db: Session = Depends(get_db), id: int = Depends(auth.get_active_user)):
     todo = models.Todos(**todo.dict())
+    todo.owner_id = id
     db.add(todo)
     db.commit()
     db.refresh(todo)
     return todo
 
 
-@router.get('/{id}/')
-def get_todo(id: int, db: Session = Depends(get_db)):
+@router.get('/{id}/', response_model=schemas.Todo)
+def get_todo(id: int, db: Session = Depends(get_db), user_id: int = Depends(auth.get_active_user)):
     todo = db.query(models.Todos).filter(models.Todos.id == id).first()
 
     if not todo:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"User with id: {id} does not exist")
+    if not todo.owner_id == user_id:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail="You do not have permission to access this resource")
+
     return todo
 
 
-@router.put('/{id}/')
-def update_todo(id: int, todo: schemas.CreateTodo, db: Session = Depends(get_db)):
+@router.put('/{id}/', response_model=schemas.Todo)
+def update_todo(id: int, todo: schemas.CreateTodo, db: Session = Depends(get_db), user_id: int = Depends(auth.get_active_user)):
     todo_query = db.query(models.Todos).filter(models.Todos.id == id)
     tododb = todo_query.first()
 
     if not tododb:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"User with id: {id} does not exist")
+    if not tododb.owner_id == user_id:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail="You do not have permission to modify this resource")
+
     todo_query.update(todo.dict(), synchronize_session=False)
     db.commit()
     db.refresh(tododb)
@@ -50,13 +61,17 @@ def update_todo(id: int, todo: schemas.CreateTodo, db: Session = Depends(get_db)
 
 
 @router.delete('/{id}/', status_code=status.HTTP_204_NO_CONTENT)
-def delete_todo(id: int, db: Session = Depends(get_db)):
+def delete_todo(id: int, db: Session = Depends(get_db), user_id: int = Depends(auth.get_active_user)):
     todo_query = db.query(models.Todos).filter(models.Todos.id == id)
     tododb = todo_query.first()
 
     if not tododb:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"User with id: {id} does not exist")
+    if not tododb.owner_id == user_id:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail="You do not have permission to delete this resource")
+
     todo_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
