+++ src/controller/user.py
@@ -1,12 +1,12 @@
 from typing import List
 from fastapi import APIRouter, HTTPException, status, Response, Depends
 from src.schema import schemas
-from passlib.context import CryptContext
 from sqlalchemy.orm import Session
 from src.model import models
 from src.model.db_conn import get_db
+from src.utils.hashed import hashed_pwd
+from src.auth import auth
 
-pwd_context = CryptContext(schemes="bcrypt")
 
 router = APIRouter(
     prefix="/api/v1/users",
@@ -22,8 +22,7 @@ def get_user(db: Session = Depends(get_db)):
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.User)
 def create_user(user: schemas.CreateUser, db: Session = Depends(get_db)):
-    hashed_pwd = pwd_context.hash(user.password)
-    user.password = hashed_pwd
+    user.password = hashed_pwd(user.password)
     usr = models.User(**user.dict())
     db.add(usr)
     db.commit()
@@ -31,8 +30,8 @@ def create_user(user: schemas.CreateUser, db: Session = Depends(get_db)):
     return usr
 
 
-@router.get('/{id}', status_code=status.HTTP_200_OK)
-def get_user(id: int, db: Session = Depends(get_db)):
+@router.get('/{id}', status_code=status.HTTP_200_OK, response_model=schemas.User)
+def get_user(id: int, db: Session = Depends(get_db), _: int = Depends(auth.get_active_user)):
     user = db.query(models.User).filter(models.User.id == id).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
@@ -41,12 +40,16 @@ def get_user(id: int, db: Session = Depends(get_db)):
 
 
 @router.put('/{id}')
-def update_user(id: int, usr: schemas.CreateUser, db: Session = Depends(get_db)):
+def update_user(id: int, usr: schemas.CreateUser, db: Session = Depends(get_db), user_id: int = Depends(auth.get_active_user)):
     user_query = db.query(models.User).filter(models.User.id == id)
     user = user_query.first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"User with id: {id} does not exist")
+    if not user.id == user_id:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail="You do not have permission to modify this resource")
+
     user_query.update(usr.dict(), synchronize_session=False)
     db.commit()
     db.refresh(user)
@@ -54,12 +57,16 @@ def update_user(id: int, usr: schemas.CreateUser, db: Session = Depends(get_db))
 
 
 @router.delete('/{id}', status_code=status.HTTP_204_NO_CONTENT)
-def delete_user(id: int, db: Session = Depends(get_db)):
+def delete_user(id: int, db: Session = Depends(get_db), user_id: int = Depends(auth.get_active_user)):
     user_query = db.query(models.User).filter(models.User.id == id)
     user = user_query.first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"User with id: {id} does not exist")
+    if not user.id == user_id:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail="You do not have permission to delete this resource")
+
     user_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
